==> notebooks/01.0-dm-extract_relevant_metadata.ipynb <==
import pandas as pd
import json
import os

# load books metadata
load_path = "../data/raw/SFGram-dataset-master/books/"

books = []

for f in os.listdir(load_path):
    if f.endswith("json"):
        with open(load_path + f) as file:
            book_data = json.load(file)

        # filter out magazines
        if not (
            book_data["title"].startswith("Galaxy")
            or book_data["title"].startswith("IF")
        ):
            # check if the fulltext actually exists in the data
            id = f[7:10]  # only need 4 digits because << 10k books
            book_fulltext_path = "../data/raw/SFGram-dataset-master/book-contents/"
            book_fulltext_path += f"book00{id}.txt"
            if os.path.isfile(book_fulltext_path):
                
                book = {
                    "id": id,  
                    "title": book_data["title"],
                    "author": book_data["author_name"],
                    "year": book_data["year"],
                }

                books.append(book)

books

df = pd.DataFrame(books)
df = df.sort_values(by="id")
# df.set_index('id', inplace=True)
df

save_path = "../data/interim/books/books_metadata.txt"
df.to_csv(save_path, index=False)
