==> notebooks/03.0-dm-topic_modelling.ipynb <==
# imports
from bertopic import BERTopic
import pandas as pd
import spacy
import json
import os

# load book metadata dataframe
df_metadata = pd.read_csv("../data/interim/books/books_metadata.txt", dtype="str")
ids = list(df_metadata["id"])

# open a fulltext
id = "003"
load_path = "../data/interim/books/fulltexts/"
book_path = load_path + f"book{id}.txt"
with open(book_path) as file:
    fulltext = file.read()

# split into satisfactory chunks using spaCy
nlp = spacy.load("en_core_web_sm")
doc = nlp(fulltext)

chunk_lists = []
token_count = 0
chunk_list = []
for sent in doc.sents:
    l = len(sent)
    if token_count + l > 400:
        # 400 is an arbitrary value, keeping a large buffer
        # to keep BERT's tokens below 512
        chunk_lists.append(chunk_list)
        token_count = l
        chunk_list = []
        chunk_list.append(sent.text)
    else:
        token_count += l
        chunk_list.append(sent.text)
chunks = [" ".join(chunk_list) for chunk_list in chunk_lists]

# run BERTopic
model = BERTopic.load("MaartenGr/BERTopic_Wikipedia")
df_fulltopic = model.get_topic_info()

topics, probabilities = model.transform(chunks)

# extract topics and probabililties
topic_dicts = {}

for topic in topics:
    topic_dicts[topic] = {
        "prob": 0,
        "words": list(
            df_fulltopic.loc[df_fulltopic["Topic"] == topic]["Representation"]
        ),
    }

for i, topic in enumerate(topics):
    topic_dicts[topic]["prob"] += probabilities[i]

df_topic = pd.DataFrame.from_dict(topic_dicts, orient="index")
df_topic = df_topic.sort_values(by = 'prob', ascending=False)
df_topic

# full pipeline

failed_ids_tm = []

# directories
load_path = "../data/interim/books/fulltexts/"
save_path = "../data/interim/topics/"

# models
nlp = spacy.load("en_core_web_sm")
model = BERTopic.load("MaartenGr/BERTopic_Wikipedia")
df_fulltopic = model.get_topic_info()

for id in ids[0:]:
    try:
        # open the fulltext
        book_path = load_path + f"book{id}.txt"
        with open(book_path) as file:
            fulltext = file.read()

        # split into satisfactory chunks using spaCy
        doc = nlp(fulltext)

        chunk_lists = []
        token_count = 0
        chunk_list = []
        for sent in doc.sents:
            l = len(sent)
            if token_count + l > 400:
                # 400 is an arbitrary value, keeping a large buffer
                # to keep BERT's tokens below 512
                chunk_lists.append(chunk_list)
                token_count = l
                chunk_list = []
                chunk_list.append(sent.text)
            else:
                token_count += l
                chunk_list.append(sent.text)
        chunks = [" ".join(chunk_list) for chunk_list in chunk_lists]

        # run topic modelling inference on each chunk
        topics, probabilities = model.transform(chunks)

        # extract topics and probabililties
        topic_dicts = {}

        for topic in topics:
            topic_dicts[topic] = {
                "prob": 0,
                "words": list(
                    df_fulltopic.loc[df_fulltopic["Topic"] == topic]["Representation"]
                ),
            }

        for i, topic in enumerate(topics):
            topic_dicts[topic]["prob"] += probabilities[i]

        df_topic = pd.DataFrame.from_dict(topic_dicts, orient="index")
        df_topic = df_topic.sort_values(by = 'prob', ascending=False)

        # write topic_df to file
        out_file = save_path + f"{id}.csv"
        df_topic.to_csv(out_file, index=False)
        print(f"ID:{id} done.")

    except:
        failed_ids_tm.append(id)
        print(f"ID:{id} failed.")
