==> notebooks/04.0-dm-remove_bad_books.ipynb <==
Some books weren't handled correctly by the emotion detector and/or the topic modeller. \
For simplicity, we'll use this file to remove those.

import os

emotion_path = "../data/interim/emotions/"
topic_path = "../data/interim/topics/"

em_ids, top_ids = [], []

for f in os.listdir(emotion_path):
    if f.endswith("txt"):
        em_ids.append(f[:3])

for f in os.listdir(topic_path):
    if f.endswith("csv"):
        top_ids.append(f[:3])

em_set, top_set = set(em_ids), set(top_ids)

bad_em_ids = []
for id in em_ids:
    if id not in top_set:
        bad_em_ids.append(id)

bad_top_ids = []
for id in top_ids:
    if id not in em_set:
        bad_top_ids.append(id)

for id in bad_em_ids:
    f = emotion_path + id + ".txt"
    if os.path.exists(f):
        os.remove(f)

for id in bad_top_ids:
    f = topic_path + id + ".csv"
    if os.path.exists(f):
        os.remove(f)

bad_em_ids
