==> notebooks/04.1-dm-collect_emotions.ipynb <==
import json
import os

load_path = "../data/interim/emotions/"
save_path = "../data/final/"

emotions = {
    "joy": {},
    "love": {},
    "anger": {},
    "sadness": {},
    "fear": {},
    "surprise": {},
}

ids = []
for f in os.listdir(load_path):
    ids.append(f[:3])
ids.sort()

for id in ids:
    f = id + ".txt"
    with open(load_path + f) as file:
        ems = json.load(file)

    # normalize emotion scores
    tot = sum([ems[em] for em in ems])
    for em in ems:
        ems[em] /= tot

    # save them to the dict
    for em in emotions:
        emotions[em][id] = ems[em]

# write emotion dict
out_file = save_path + "emotions.json"
with open(out_file, "w") as file:
    json.dump(emotions, file)

ems

len(emotions["joy"])
