==> notebooks/04.3-dm-collect_misc_book_data.ipynb <==
import pandas as pd
import json
import os

load_path_metadata = "../data/raw/SFGram-dataset-master/books/"
load_path_interim = "../data/interim/"
save_path = "../data/final/"

books = {}

ids = []
for f in os.listdir(load_path_interim + "emotions/"):
    ids.append(f[:3])
ids.sort()

for id in ids:
    # load book metadata
    f_book = "books00" + id + ".json"
    with open(load_path_metadata + f_book) as file:
        book = json.load(file)

    # load book emotions
    f_ems = "emotions/" + id + ".txt"
    with open(load_path_interim + f_ems) as file:
        ems = json.load(file)
    book["emotion_scores"] = ems

    # load book topics
    f_tops = "topics/" + id + ".csv"
    df_topics = pd.read_csv(load_path_interim + f_tops)
    # normalize probabilities
    tot = sum(df_topics["prob"])
    df_topics["prob"] /= tot
    
    book["topic_words"] = df_topics.values.tolist()

    # save updated metadata
    books[id] = book

out_file = save_path + "books.json"
with open(out_file, "w") as file:
    json.dump(books, file)
