==> notebooks/05.1-dm-plotly_plots.ipynb <==
import plotly.express as px
import pandas as pd
import json
import os

# load in data

load_path = "../data/final/"

with open(load_path + "emotions.json") as f:
    emotions = json.load(f)
with open(load_path + "topics.json") as f:
    topics = json.load(f)
with open(load_path + "vocab.json") as f:
    vocab = json.load(f)
with open(load_path + "books.json") as f:
    books = json.load(f)

# query 1: given an emotion and the topic word,
# return books sorted in descending order wrt
# the product of the emotion and topic word's scores

# create a horizontal bar chart of the relevant books

emotion = "surprise"
word = "travel"

# emotions
emotions_dict = emotions[emotion]

# topic word
relevant_topics = vocab[word]

word_scores_dict_list = [topics[topic] for topic in relevant_topics]
word_score_dict = {}
for d in word_scores_dict_list:
    for k in d:
        word_score_dict[k] = word_score_dict.get(k, 0) + d[k]

# product of the two
product_dict = {}
for id in word_score_dict:
    book_name = books[id]["title"]
    product_dict[book_name] = word_score_dict[id] * emotions_dict[id]

# # don't need the below, tbh
# product_list = [[k, v] for k, v in product_dict.items()]
# product_list.sort(key=lambda x: x[1], reverse=True)

# horizontal bar chart
product_dict
df = pd.DataFrame([product_dict]).transpose()
df = df.reset_index()
df = df.rename(columns={"index": "book_name", 0: "score"})
df = df.sort_values(by="score", ascending=False)
df = df.head(10)

fig = px.bar(df, x="score", y="book_name", orientation="h")
fig.update_layout(yaxis={"categoryorder": "total ascending"})
# fig.update_yaxes(ticktext=df["small_topic"], tickvals=df[1])

fig.show()

# query 2: given a book, create a horizontal bar chart of that book's topic words
book_id = "001"
topic_probs_words = books[book_id]["topic_words"]
df = pd.DataFrame(topic_probs_words)
df = df.rename(columns={0: "score", 1: "topic_words"})
df.sort_values(by="score")
df = df.head(10)
df["small_topic"] = df["topic_words"].apply(lambda x: x[:75] + " ... ")

fig = px.bar(df, x="score", y="topic_words", orientation="h")
fig.update_layout(yaxis={"categoryorder": "total ascending"})
fig.update_yaxes(ticktext=df["small_topic"], tickvals=df["topic_words"])

fig.show()

df = pd.DataFrame(topic_probs_words)

type(df[0][0])

vocab['agriculture']

# Sample data creation
topic_probs_words = {
    "Topic": ["Topic A", "Topic B", "Topic C", "Topic D", "Topic E"],
    "Probability": [0.2, 0.4, 0.1, 0.3, 0.5],
}

df = pd.DataFrame(topic_probs_words)

# Set the maximum length of y-axis labels
max_label_length = 4
df["TruncatedTopic"] = df["Topic"].apply(lambda x: x[:max_label_length])

fig = px.bar(df, x="Probability", y="TruncatedTopic", orientation="h")

# Update y-axis labels
fig.update_yaxes(
    ticktext=df["TruncatedTopic"], tickvals=df["Topic"], title_text="Topic"
)

fig.show()

# emotion plot for a book
book_id = "001"
book_emotions = books[book_id]["emotion_scores"]
df = pd.DataFrame([book_emotions])
# df = df.reset_index()

# df = df.rename(columns={0: "score", 1: "topic_words"})
# df.sort_values(by="score")
# df = df.head(10)
# df["small_topic"] = df["topic_words"].apply(lambda x: x[:75] + " ... ")

df

fig = px.bar(df, orientation="h",
             labels=dict(
                 value="Score %age", index="", variable="Emotion"
             ))
fig.update_xaxes(range=[0,100])
fig.update_yaxes(showticklabels=False)
# fig.update_layout(yaxis={"categoryorder": "total ascending"})
# fig.update_yaxes(ticktext=df["small_topic"], tickvals=df["topic_words"])

fig.show()

df = df.transpose()
tot = sum(df[0])
df[0] = df[0] * 100 / tot
df = df.transpose()
df
